Added the new Help & Feedback popover to the account side menu
